refactor: split auth_verify.rs and integration_logout_tests.rs under 600-line cap
- Extract verify_email handler + VerifyEmailRequest to routes/auth_verify.rs
- Extract integration_logout_full_flow to tests/integration_logout_tests.rs
- Update mod.rs and openapi.rs to reference new auth_verify module
- All source files now under 600-line CI-enforced limit
