+++ services/control-api/src/routes/mod.rs
@@ -1,6 +1,7 @@
 pub mod api_keys;
 pub mod auth;
 pub mod auth_logout;
+pub mod auth_verify;
 pub mod health;
 pub mod me;
 pub mod tenants;
@@ -9,8 +10,9 @@ use axum::{Router, routing::{delete, get, post, put}};
 
 use crate::routes::{
     api_keys::{create_api_key, list_api_keys, revoke_api_key},
-    auth::{forgot_password, login, reset_password, signup, verify_email},
+    auth::{forgot_password, login, reset_password, signup},
     auth_logout::logout,
+    auth_verify::verify_email,
     health::{health_check, openapi_json, ready_check},
     me::switch_tenant,
     tenants::{invite_member, remove_member, transfer_ownership, update_member_role},
